Add q0, q1, q2
