+++ tests/FlowtideDotNet.Nexmark/Internal/Builders/AuctionBatchBuilder.cs
@@ -7,9 +7,9 @@ using System.Collections.Generic;
 using System.IO;
 using System.IO.Pipelines;
 
-namespace FlowtideDotNet.Nexmark.Internal;
+namespace FlowtideDotNet.Nexmark.Internal.Builders;
 
-internal sealed class PersonBatchBuilder
+internal sealed class AuctionBatchBuilder
 {
     private readonly IMemoryAllocator _memoryAllocator;
     private readonly int _batchSize;
@@ -20,26 +20,28 @@ internal sealed class PersonBatchBuilder
 
     public int EventCount { get; private set; }
 
-    public PersonBatchBuilder(IMemoryAllocator memoryAllocator, int batchSize)
+    public AuctionBatchBuilder(IMemoryAllocator memoryAllocator, int batchSize)
     {
         _memoryAllocator = memoryAllocator;
         _batchSize = batchSize;
         _currentColumns = CreateColumns();
         _currentRowCount = 0;
         _eventBatchSerializer = new EventBatchSerializer();
-        _fileStream = File.OpenWrite("person_batches.bin");
+        _fileStream = File.OpenWrite("auction_batches.bin");
     }
 
-    public void Add(in Person person)
+    public void Add(in Auction auction)
     {
-        _currentColumns[0].Add(new Int64Value(person.Id));
-        _currentColumns[1].Add(new StringValue(person.Name));
-        _currentColumns[2].Add(new StringValue(person.EmailAddress));
-        _currentColumns[3].Add(new StringValue(person.CreditCard));
-        _currentColumns[4].Add(new StringValue(person.City));
-        _currentColumns[5].Add(new StringValue(person.State));
-        _currentColumns[6].Add(new StringValue(person.DateTime));
-        _currentColumns[7].Add(new StringValue(person.Extra));
+        _currentColumns[0].Add(new Int64Value(auction.Id));
+        _currentColumns[1].Add(new StringValue(auction.ItemName));
+        _currentColumns[2].Add(new StringValue(auction.Description));
+        _currentColumns[3].Add(new Int64Value(auction.InitialBid));
+        _currentColumns[4].Add(new Int64Value(auction.Reserve));
+        _currentColumns[5].Add(new StringValue(auction.DateTime));
+        _currentColumns[6].Add(new StringValue(auction.Expires));
+        _currentColumns[7].Add(new Int64Value(auction.Seller));
+        _currentColumns[8].Add(new Int64Value(auction.Category));
+        _currentColumns[9].Add(new StringValue(auction.Extra));
 
         _currentRowCount++;
         EventCount++;
@@ -60,7 +62,7 @@ internal sealed class PersonBatchBuilder
             
             var span = bufferWriter.WrittenSpan;
             _fileStream.Write(span);
-            
+
             _currentColumns = CreateColumns();
             _currentRowCount = 0;
         }
@@ -74,8 +76,8 @@ internal sealed class PersonBatchBuilder
 
     private IColumn[] CreateColumns()
     {
-        var columns = new IColumn[8];
-        for (int i = 0; i < 8; i++)
+        var columns = new IColumn[10];
+        for (int i = 0; i < 10; i++)
         {
             columns[i] = Column.Create(_memoryAllocator);
         }
