+++ tests/FlowtideDotNet.Nexmark/Internal/Builders/BidBatchBuilder.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.IO.Pipelines;
 
-namespace FlowtideDotNet.Nexmark.Internal;
+namespace FlowtideDotNet.Nexmark.Internal.Builders;
 
 internal sealed class BidBatchBuilder
 {
