+++ tests/FlowtideDotNet.Nexmark/QueryBase.cs
@@ -1,0 +1,48 @@
+﻿// Licensed under the Apache License, Version 2.0 (the "License")
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//  
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using BenchmarkDotNet.Attributes;
+using FlowtideDotNet.Nexmark.Internal;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FlowtideDotNet.Nexmark
+{
+    public class QueryBase
+    {
+        private int iterationId = 0;
+        private NexmarkQueryStream _stream = null!;
+        private NexmarkGenerator _generator = null!;
+        private NexmarkDataStream _dataStream = null!;
+
+        protected NexmarkQueryStream Stream => _stream;
+
+        [GlobalSetup]
+        public void Setup()
+        {
+            FileDataGenerator.GenerateData(10_000_000, 10_000);
+        }
+
+        [IterationSetup]
+        public void IterationSetup()
+        {
+            _stream = new NexmarkQueryStream(_dataStream, iterationId.ToString())
+            {
+                CachePageCount = 100_000
+            };
+            iterationId++;
+        }
+    }
+}
