--- tests/FlowtideDotNet.Nexmark/nuget.config
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-	<packageSources>
-		<clear />
-		<add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
-		<add key="bdn-nightly" value="https://www.myget.org/F/benchmarkdotnet/api/v3/index.json" />
-	</packageSources>
-</configuration>
